requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
ipdb==0.8
ipython==1.1.0
itsdangerous==0.23
pytz==2013.8
requests==2.1.0
whenpy==0.3.0
wsgiref==0.1.2